+++ client/src/pages/deploy/deployStyles.ts
@@ -1,0 +1,2 @@
+/** Muted hint text used across the deploy wizard steps. */
+export const HINT_STYLE: React.CSSProperties = { fontSize: 12, color: 'var(--muted)' }
